+++ naive.py
@@ -1,26 +1,8 @@
 import time
-
 from queue import PriorityQueue
+
 from input import read_test_cases, MATRIX_SIZE
-
-
-def calculate_permutations(matrix):
-    options = [(-1, 0, "L"), (1, 0, "R"), (0, -1, "U"), (0, 1, "D")]
-    permutations = []
-    row = col = -1
-    for i in range(MATRIX_SIZE):
-        for j in range(MATRIX_SIZE):
-            if matrix[i][j] == 0:
-                row = i
-                col = j
-
-    for i, j, letter in options:
-        if 0 <= row + i < MATRIX_SIZE and 0 <= col + j < MATRIX_SIZE:
-            temp = matrix
-            temp[row][col], temp[i][j] = temp[i][j], temp[row][col]
-            permutations.append((temp, letter))
-
-    return permutations
+from utils import calculate_permutations, check_answer
 
 
 def calculate_heuristic(matrix):
@@ -33,11 +15,6 @@ def calculate_heuristic(matrix):
                 correct_col = (num - 1) % MATRIX_SIZE
                 cost += abs(i - correct_row) + abs(j - correct_col)
     return cost
-
-
-# TODO
-def check_answer():
-    pass
 
 
 def main():
